warning fixed

+++ criteria1D/water1D.h
@@ -7,7 +7,7 @@
     bool computeInfiltration(Criteria1D* myCase, float prec, float sprayIrrigation);
     bool computeSurfaceRunoff(Criteria1D* myCase);
     bool computeLateralDrainage(Criteria1D* myCase);
-    bool computeCapillaryRise(Criteria1D* myCase, float waterTableDepth);
+    bool computeCapillaryRise(Criteria1D* myCase, double waterTableDepth);
 
     double getSoilWaterContent(Criteria1D* myCase);
